chore: Remove inappropriate links and add popup notification functionality
This commit removes the inappropriate links to adult websites and adds functionality to display a popup notification when the "Click 10 cái nhận quà!" button is clicked. The popup notification asks for permission to show notifications and, if granted, displays a random notification with a body message and an icon. The code for the popup notification is added in a script tag at the end of the index.html file.